+++ src/markdown-engine/heading-id-generator.ts
@@ -22,8 +22,8 @@ export default class HeadingIdGenerator {
       .replace(/~|。/g, '') // sanitize
       .replace(/``(.+?)``\s?/g, replacement)
       .replace(/`(.*?)`\s?/g, replacement)
-      .replace(/\s__([^_]+?)__\s/g, `-$1-`)
-      .replace(/\s_([^_]+?)_\s/g, `-$1-`);
+      .replace(/(^|\s)__([^_]+?)__(\s|$)/g, `$1$2$3`)
+      .replace(/(^|\s)_([^_]+?)_(\s|$)/g, `$1$2$3`);
     let slug = uslug(heading.replace(/\s/g, '~')).replace(/~/g, '-');
     if (this.table[slug] >= 0) {
       this.table[slug] = this.table[slug] + 1;